+++ package.json
@@ -9,21 +9,17 @@
     "deploy": "next build && next export"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.2",
     "@svgr/webpack": "^6.2.1",
     "axios": "^0.26.0",
     "next": "12.0.10",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-redux": "^7.2.6",
-    "redux": "^4.1.2",
     "styled-components": "^5.3.3",
     "styled-reset": "^4.3.4"
   },
   "devDependencies": {
     "@types/node": "17.0.17",
     "@types/react": "17.0.39",
-    "@types/react-redux": "^7.1.22",
     "@types/styled-components": "^5.1.22",
     "babel-plugin-styled-components": "^2.0.2",
     "eslint": "8.9.0",
