+++ src/types/clearance.ts
@@ -1,0 +1,39 @@
+import { ConCategory2 } from "./category";
+
+// 땡처리 리스트 타입
+export interface ClearanceList {
+  conItems: ClearanceListConItems[];
+}
+
+export interface ClearanceListConItems {
+  id: number;
+  name: string;
+  originalPrice: number;
+  createdAt: string;
+  sfId: string;
+  minSellingPrice: number;
+  ncSellingPrice: number;
+  count: number;
+  information: any;
+  tip: any;
+  warning?: string;
+  discountRate: number;
+  askingPrice: number;
+  isRefuse: number;
+  isBlock: number;
+  info: any;
+  isOnlyAccount: number;
+  conCategory2Id: number;
+  imageUrl: string;
+  conCategory2: ConCategory2;
+}
+
+export interface ClearanceListConCategory1 {
+  id: number;
+  name: string;
+  createdAt: string;
+  priority: number;
+  discountRate: number;
+  info: string;
+  imageUrl: string;
+}
